chore: Refactor GuestLogin and Welcome pages to use custom hooks
